inserter/requirements.txt
mysql-connector-python==8.0.31
Faker==18.9.0